We need to concatenate the strings in a row together when we grab things

+++ src/datacode/feature_dict.py
@@ -49,7 +49,7 @@ class Data():
 
 	def getColumn(self, var):
 		ranges = self.lookUp(var = var)[1]
-		return self.data[:,ranges[0]:ranges[1]]
+		return self.data[:,ranges[0]:ranges[1]].astype('float')
 
 	def save(self, filename):
 		with open(filename, 'wb') as f:
